Minor cleanup of claude's code
- Add vim gitignore
- Move pi0_fuse config into its own file, keeping
  the same style as surrounding files
- Remove some unused variables in training config
