+++ src/openpi/models/pi0_fuse_config.py
@@ -1,0 +1,91 @@
+import dataclasses
+from typing import TYPE_CHECKING
+
+import flax.nnx as nnx
+import jax
+import jax.numpy as jnp
+from typing_extensions import override
+
+from openpi.models import model as _model
+import openpi.models.gemma as _gemma
+from openpi.shared import array_typing as at
+import openpi.shared.nnx_utils as nnx_utils
+
+if TYPE_CHECKING:
+    from openpi.models.pi0_fuse import Pi0Fuse
+
+@dataclasses.dataclass(frozen=True)
+class Pi0FuseConfig(_model.BaseModelConfig):
+    """Config for Pi0Fuse model (pi05-compatible with reasoning loss)."""
+    dtype: str = "bfloat16"
+    paligemma_variant: _gemma.Variant = "gemma_2b"
+    action_expert_variant: _gemma.Variant = "gemma_300m"
+
+    action_dim: int = 32
+    action_horizon: int = 16
+    max_token_len: int = 415
+
+    # L = L_text + diffusion_loss_coeff * L_diffusion
+    diffusion_loss_coeff: float = 1.0
+
+    # Pi05 uses AdaRMSNorm in action expert and discretized state in text
+    pi05: bool = True
+
+    @property
+    @override
+    def model_type(self) -> _model.ModelType:
+        return _model.ModelType.PI0_FUSE
+
+    @override
+    def create(self, rng: at.KeyArrayLike) -> "Pi0Fuse":
+        from openpi.models.pi0_fuse import Pi0Fuse
+        return Pi0Fuse(self, rngs=nnx.Rngs(rng))
+
+    @override
+    def inputs_spec(self, *, batch_size: int = 1) -> tuple[_model.FuseObservation, _model.Actions]:
+        image_spec = jax.ShapeDtypeStruct([batch_size, *_model.IMAGE_RESOLUTION, 3], jnp.float32)
+        image_mask_spec = jax.ShapeDtypeStruct([batch_size], jnp.bool_)
+
+        with at.disable_typechecking():
+            observation_spec = _model.FuseObservation(
+                images={
+                    "base_0_rgb": image_spec,
+                    "left_wrist_0_rgb": image_spec,
+                    "right_wrist_0_rgb": image_spec,
+                },
+                image_masks={
+                    "base_0_rgb": image_mask_spec,
+                    "left_wrist_0_rgb": image_mask_spec,
+                    "right_wrist_0_rgb": image_mask_spec,
+                },
+                state=jax.ShapeDtypeStruct([batch_size, self.action_dim], jnp.float32),
+                tokenized_prompt=jax.ShapeDtypeStruct([batch_size, self.max_token_len], jnp.int32),
+                tokenized_prompt_mask=jax.ShapeDtypeStruct([batch_size, self.max_token_len], jnp.bool_),
+                token_ar_mask=jax.ShapeDtypeStruct([batch_size, self.max_token_len], jnp.int32),
+                token_loss_mask=jax.ShapeDtypeStruct([batch_size, self.max_token_len], jnp.bool_),
+                diffusion_loss_mask=jax.ShapeDtypeStruct([batch_size], jnp.bool_),
+            )
+        action_spec = jax.ShapeDtypeStruct([batch_size, self.action_horizon, self.action_dim], jnp.float32)
+        return observation_spec, action_spec
+
+    def get_freeze_filter(self) -> nnx.filterlib.Filter:
+        filters = []
+        has_lora = False
+        gemma_params_filter = nnx_utils.PathRegex(".*llm.*")
+        action_expert_params_filter = nnx_utils.PathRegex(".*llm.*_1.*")
+        if "lora" in self.paligemma_variant:
+            filters.append(gemma_params_filter)
+            if "lora" not in self.action_expert_variant:
+                filters.append(nnx.Not(action_expert_params_filter))
+            has_lora = True
+        elif "lora" in self.action_expert_variant:
+            filters.append(action_expert_params_filter)
+            has_lora = True
+
+        if has_lora:
+            filters.append(nnx.Not(nnx_utils.PathRegex(".*lora.*")))
+        if not filters:
+            return nnx.Nothing
+        return nnx.All(*filters)
+
+
